Rename internal lib
